--- frontend/src/views/Collection/components/EmptyDatasets/index.tsx
@@ -1,46 +1,0 @@
-import { Button, H4, Intent, UL } from "@blueprintjs/core";
-import React, { FC } from "react";
-import DropboxChooser, {
-  Props as DropboxChooserProps,
-} from "src/components/DropboxChooser";
-import { StyledLink } from "src/views/Collection/common/style";
-import { Border, CenterAlignedDiv } from "./style";
-
-const CLI_README_LINK =
-  "https://github.com/chanzuckerberg/cellxgene/blob/main/dev_docs/schema_guide.md";
-
-interface Props {
-  onUploadFile: DropboxChooserProps["onUploadFile"];
-}
-
-const EmptyDatasets: FC<Props> = ({ onUploadFile }) => {
-  return (
-    <Border>
-      <CenterAlignedDiv>
-        <H4>No datasets uploaded</H4>
-        <div>
-          Before you begin uploading dataset files:
-          <UL>
-            <li>
-              You must validate your dataset locally. We provide a local CLI
-              script to do this.{" "}
-              <StyledLink href={CLI_README_LINK}>Learn More</StyledLink>
-            </li>
-            <li>
-              We only support adding datasets in the h5ad format at this time.
-            </li>
-          </UL>
-        </div>
-        <DropboxChooser onUploadFile={onUploadFile}>
-          <Button
-            intent={Intent.PRIMARY}
-            outlined
-            text={"Add Dataset from Dropbox"}
-          />
-        </DropboxChooser>
-      </CenterAlignedDiv>
-    </Border>
-  );
-};
-
-export default EmptyDatasets;
